+++ src/app/screens/Setup/SetupScreen.tsx
@@ -1,1 +1,0 @@
-export { SetupScreen, default } from "./SetupScreenCore";